+++ genomepy/annotation/mygene.py
@@ -7,7 +7,7 @@ from loguru import logger
 from tqdm import tqdm
 
 from genomepy.annotation.utils import _parse_annot
-from genomepy.caching import memory
+from genomepy.caching import cache_exp_short, disk_cache
 
 
 def _map_genes(
@@ -58,7 +58,7 @@ def _map_genes(
     # remove version numbers from gene IDs
     split_id = df["name"].str.split(r"\.", expand=True)[0]
     df = df.assign(split_id=split_id.values)
-    genes = set(split_id)
+    genes = sorted(set(split_id))
 
     result = query_mygene(genes, self.tax_id, field)
     # result = _query_mygene(self, genes, field=to)
@@ -78,7 +78,7 @@ def _map_genes(
     return df
 
 
-@memory.cache
+@disk_cache.memoize(expire=cache_exp_short, tag="query_mygene")
 def query_mygene(
     query: Iterable[str],
     tax_id: Union[str, int],