+++ genomepy/providers/local.py
@@ -36,12 +36,12 @@ class LocalProvider(BaseProvider):
     def assembly_accession(self, name):
         return
 
-    def search(self, term):
+    def search(self, term, size=False):
         """return an empty generator,
         same as if no genomes were found at the other providers"""
         yield from ()
 
-    def _genome_info_tuple(self, name):
+    def _genome_info_tuple(self, name, size=False):
         return tuple()
 
     def _check_name(self, name):
@@ -107,7 +107,7 @@ class LocalProvider(BaseProvider):
         hits = []
         for ext in ["gtf", "gff", "gff3"]:
             # .*? = non greedy filler. (\.gz)? = optional .gz
-            expr = f"{search_name}.*?\.{ext}(\.gz)?"  # noqa: W605
+            expr = fr"{search_name}.*?\.{ext}(\.gz)?"  # noqa: W605
             for line in search_list:
                 hit = re.search(expr, line, flags=re.IGNORECASE)
                 if hit: