+++ tests/test_09_provider_ensembl.py
@@ -10,9 +10,9 @@ def test_ensemblprovider(ensembl):
 
 
 def test_genome_info_tuple(ensembl):
-    t = ensembl._genome_info_tuple("KH")
+    t = ensembl._genome_info_tuple("KH", size=True)
     assert isinstance(t, tuple)
-    assert t[0:3] == ("KH", "GCA_000224145.1", 7719)
+    assert t[:-1] == ("KH", "GCA_000224145.1", 7719, True, 'Ciona intestinalis', 115227500)
 
 
 def test_get_version(ensembl):